refactor(desktop): 连接运行时从 index.ts 抽到 connections-runtime
把启动序列里的连接接线 / ping / 热重配（connectionRuntime + wireConnections / pingConnections /
reconfigureConnections / persistConnectionUser / activeConnectionIds，~95 行闭包 + 连接状态）抽到
createConnectionRuntime（connections-runtime.ts），index.ts 改为 `const conns = createConnectionRuntime({...})`
并经 conns.wire / conns.ping / conns.reconfigure / conns.runtime / conns.activeConnectionIds 使用。
连接级本地状态内聚到工厂（随 ping 增量回写），不再是 index 模块级可变全局。纯搬运、行为不变。

desktop typecheck/lint/build 全绿。
